+++ wasm/cpu/opcodes.ts
@@ -75,131 +75,12 @@ import {
   updateSound
 } from '../sound/index';
 
-// Public funciton to run opcodes until an event occurs.
-// Return values:
-// -1 = error
-// 1 = render a frame
-// 2 = replace boot rom
-export function update(): i32 {
-
-  let error: boolean = false;
-  let numberOfCycles: i32 = -1;
-
-  while(!error &&
-    Cpu.currentCycles < Cpu.MAX_CYCLES_PER_FRAME()) {
-    numberOfCycles = emulationStep();
-    if (numberOfCycles < 0) {
-      error = true;
-    }
-  }
-
-  // Find our exit reason
-  if (Cpu.currentCycles >= Cpu.MAX_CYCLES_PER_FRAME()) {
-    // Render a frame
-
-    // Reset our currentCycles
-    Cpu.currentCycles -= Cpu.MAX_CYCLES_PER_FRAME();
-
-    return 1;
-  }
-  // TODO: Boot ROM handling
-
-  // There was an error, return -1, and push the program counter back to grab the error opcode
-  Cpu.programCounter -= 1;
-  return -1;
-}
-
-// Function to execute an opcode, and update other gameboy hardware.
-// http://www.codeslinger.co.uk/pages/projects/gameboy/beginning.html
-export function emulationStep(): i32 {
-  // Get the opcode, and additional bytes to be handled
-  // Number of cycles defaults to 4, because while we're halted, we run 4 cycles (according to matt :))
-  let numberOfCycles: i32 = 4;
-  let opcode: i32 = 0;
-
-  // Cpu Halting best explained: https://www.reddit.com/r/EmuDev/comments/5ie3k7/infinite_loop_trying_to_pass_blarggs_interrupt/db7xnbe/
-  if(!Cpu.isHalted && !Cpu.isStopped) {
-    opcode = <u8>eightBitLoadFromGBMemory(Cpu.programCounter);
-    numberOfCycles = executeOpcode(opcode);
-  } else {
-    // if we were halted, and interrupts were disabled but interrupts are pending, stop waiting
-    if(Cpu.isHalted && !Interrupts.masterInterruptSwitch && Interrupts.areInterruptsPending()) {
-      Cpu.isHalted = false;
-      Cpu.isStopped = false;
-
-      // Need to run the next opcode twice, it's a bug menitoned in
-      // The reddit comment mentioned above
-
-      // CTRL+F "low-power" on gameboy cpu manual
-      // http://marc.rawer.de/Gameboy/Docs/GBCPUman.pdf
-      // E.g
-      // 0x76 - halt
-      // FA 34 12 - ld a,(1234)
-      // Becomes
-      // FA FA 34 ld a,(34FA)
-      // 12 ld (de),a
-      opcode = <u8>eightBitLoadFromGBMemory(Cpu.programCounter);
-      numberOfCycles = executeOpcode(opcode);
-      Cpu.programCounter -= 1;
-    }
-  }
-
-  // blarggFixes, don't allow register F to have the bottom nibble
-  Cpu.registerF = Cpu.registerF & 0xF0;
-
-  // Check if there was an error decoding the opcode
-  if(numberOfCycles <= 0) {
-    return numberOfCycles;
-  }
-
-  // Check if we did a DMA TRansfer, if we did add the cycles
-  if (Memory.DMACycles > 0) {
-    numberOfCycles += Memory.DMACycles;
-    Memory.DMACycles = 0;
-  }
-
-  // Interrupt Handling requires 20 cycles
-  // https://github.com/Gekkio/mooneye-gb/blob/master/docs/accuracy.markdown#what-is-the-exact-timing-of-cpu-servicing-an-interrupt
-  numberOfCycles += checkInterrupts();
-
-  // Finally, Add our number of cycles to the CPU Cycles
-  Cpu.currentCycles += numberOfCycles;
-
-  if(!Cpu.isStopped) {
-    if(Config.graphicsBatchProcessing) {
-      // Need to do this, since a lot of things depend on the scanline
-      // Batch processing will simply return if the number of cycles is too low
-      Graphics.currentCycles += numberOfCycles;
-      batchProcessGraphics();
-    } else {
-      updateGraphics(numberOfCycles);
-    }
-
-    if(Config.audioBatchProcessing) {
-      Sound.currentCycles += numberOfCycles;
-    } else {
-      updateSound(numberOfCycles);
-    }
-  }
-
-  // Check other Gameboy components
-  if (Config.timersBatchProcessing) {
-    // Batch processing will simply return if the number of cycles is too low
-    Timers.currentCycles += numberOfCycles;
-    batchProcessTimers();
-  } else {
-    updateTimers(numberOfCycles);
-  }
-
-  return numberOfCycles;
-}
-
 // Take in any opcode, and decode it, and return the number of cycles
 // Program counter can be gotten from getProgramCounter();
 // Setting return value to i32 instead of u16, as we want to return a negative number on error
 // https://rednex.github.io/rgbds/gbz80.7.html
 // http://pastraiser.com/cpu/gameboy/gameboyopcodes.html
-function executeOpcode(opcode: i32): i32 {
+export function executeOpcode(opcode: i32): i32 {
 
   // Initialize our number of cycles
   // Return -1 if no opcode was found, representing an error
